refactor(user): 修改冷却时间处理逻辑- 将冷却时间类型从 double 改为 long - 移除静态冷却计数器和相关任务 - 使用 System.currentTimeMillis() 替代自定义计数器 - 更新冷却时间设置与获取方法实现 - 注释掉不再使用的冷却处理任务调用

+++ MiniGamesBox API/src/main/java/plugily/projects/minigamesbox/api/user/IUser.java
@@ -46,7 +46,7 @@ public interface IUser {
 
   boolean checkCanCastCooldownAndMessage(String cooldown);
 
-  void setCooldown(String key, double seconds);
+  void setCooldown(String key, long seconds);
 
   double getCooldown(String key);
